requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
scikit-learn==1.3.0
